+++ src/pages/Chats/components/parseMsg/index.module.less
@@ -1,0 +1,233 @@
+:global {
+  .ogo-popover-content {
+    .ogo-popover-arrow {
+      display: none;
+    }
+    .ogo-popover-inner {
+      .ogo-popover-inner-content {
+        display: flex;
+        flex-direction: column;
+      }
+    }
+  }
+}
+.chart_content {
+  flex-grow: 1;
+  overflow-y: auto;
+}
+.group_content_wrap {
+  padding: 20px;
+  transition: all 0.7s;
+  .chats_space_Time {
+    margin: 0 auto;
+    text-align: center;
+
+    span {
+      border-radius: 4px;
+      font-size: 12px;
+      display: inline-block;
+      border-radius: 4px;
+      font-size: 12px;
+      padding: 2px 8px;
+      color: #a8abb2;
+    }
+  }
+
+  .user_head_img_wrap {
+    margin-right: 0;
+  }
+
+  .history_more {
+    text-align: center;
+    color: var(--el-color-primary);
+  }
+
+  .con_body {
+    max-width: 40vw;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+  }
+
+  .con {
+    display: flex;
+    flex-direction: row;
+    margin: 20px 0;
+    .con_img {
+      height: 35px;
+      width: 35px;
+    }
+    img {
+      max-width: 100%;
+      max-height: 400px;
+    }
+    .con_content {
+      max-width: 50%;
+      .name {
+        font-size: 12px;
+        padding-left: 4px;
+        color: #888;
+      }
+      .con_content_txt {
+        cursor: pointer;
+        text-align: left;
+        min-height: 30px;
+        padding: 7px 10px;
+        margin: 0px 0px;
+        color: black;
+        border-radius: 6px;
+        z-index: 1;
+        font-size: small;
+        word-wrap: break-word;
+        word-break: normal;
+        .emoji {
+          width: 20px;
+          height: 20px;
+          margin: 5px;
+        }
+      }
+
+      .con_content_cite_txt {
+        cursor: pointer;
+        text-align: left;
+        min-height: 30px;
+        padding: 7px 10px;
+        margin-top: 4px;
+        color: black;
+        border-radius: 6px;
+        z-index: 1;
+        max-width: 300px;
+        word-break: normal;
+        font-size: 12px;
+        color: #8a8a8a;
+        box-shadow: inset 0 0 2px #dddddd;
+        background-color: #ebebeb;
+        .emoji {
+          width: 20px;
+          height: 20px;
+          margin: 5px;
+        }
+      }
+    }
+  }
+
+  .recall {
+    font-size: 12px;
+    justify-content: center;
+
+    .reWrite {
+      margin-left: 4px;
+      cursor: pointer;
+      color: #3e5ed8;
+    }
+  }
+
+  .group_content_left {
+    .con_content {
+      max-width: 100%;
+      overflow: hidden;
+      flex-direction: column;
+      justify-content: flex-end;
+      .con_content_txt {
+        background-color: white;
+        max-width: 100%;
+        box-shadow: inset 0 0 2px #b0b0b0;
+      }
+    }
+  }
+
+  .group_content_right {
+    justify-content: flex-end;
+    .con_content {
+      max-width: 100%;
+      margin-top: 10px;
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-end;
+      align-items: flex-end;
+      .con_content_txt {
+        background-color: #b4ccf9;
+        margin-right: 10px;
+        max-width: 100%;
+        box-shadow: inset 0 0 2px #5959ff;
+      }
+      .information {
+        padding-right: 10px;
+        cursor: pointer;
+        color: #3e5ed8;
+        transform: scale(0.8);
+        font-weight: 700;
+      }
+      .readed {
+        color: #888;
+      }
+    }
+  }
+
+  .context_text_wrap {
+    position: absolute;
+    // background-color: var(--el-bg-color-overlay);
+    width: 80px;
+    height: max-content;
+    border: 1px solid var(--el-box-shadow-lighter);
+    z-index: 999;
+
+    .context_menu_item {
+      padding: 4px 6px;
+      cursor: pointer;
+      text-align: center;
+
+      &:hover {
+        // background-color: var(--el-bg-color-page);
+      }
+    }
+  }
+}
+.moreInfo {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  flex-wrap: nowrap;
+  align-items: flex-start;
+
+  li {
+    padding: 10px 2px;
+  }
+}
+
+.operate {
+  display: flex;
+  -webkit-box-orient: horizontal;
+  -webkit-box-direction: normal;
+  -ms-flex-direction: row;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  align-items: flex-start;
+  align-content: flex-start;
+}
+
+.voiceStyle {
+  audio {
+    width: 200px;
+    height: 40px;
+    overflow-x: hidden;
+  }
+
+  audio::-webkit-media-controls-start-playback-button,
+  audio::-webkit-media-controls-end-playback-button,
+  audio::-webkit-media-controls-mute-button,
+  audio::-webkit-media-controls-timeline,
+  audio::-webkit-media-controls-volume-slider,
+  audio::-webkit-media-controls-overflow-button {
+    display: none !important;
+  }
+  /* 自定义时间显示样式 */
+  audio.time-display {
+    color: #666;
+    font-size: 10px;
+  }
+}
